🔄 general OptionsComp, with CheckOptions only defining checkbox selection
